OnePathRule instance for Goal should make sense now
